Move dir for model
