+++ README.md
@@ -163,7 +163,8 @@ $ go mod download
 - infra
   - sqlhandler: DBとのコネクションを確立する
   - その他: マイグレーションスクリプトを配置する
-- model: データベースに保管されたデータをプログラム上においてどのようなデータ構造で利用するかを記述する
+- domain:
+  - model: ドメインモデルを定。データベースに保管されたカラムをプログラム上においてどのようなデータ構造で利用するかを記述する。モデルに紐付いたビジネスロジックも対象となる。
 - repository: データストアからデータを取得するための処理を記述する。クエリ結果をモデルに(GoのStruct)にマッピングし、モデルを返す
 - handler
   - router: リクエストとハンドラのマッピングをする(ルーティングの役割)
