+++ api/repository/article_repositiory.go
@@ -4,8 +4,7 @@ import (
 	"database/sql"
 	"math"
 	"time"
-
-	"yuki0920/go-blog/model"
+	"yuki0920/go-blog/domain/model"
 )
 
 func ArticleListByCursor(cursor int) ([]*model.Article, error) {
